+++ install_boot.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+DISKIMG=$1
+
+c1541 $DISKIMG -bwrite boot/bootsect.128 1 0 -bwrite boot/autostart64.128 1 1 
+c1541 $DISKIMG -@ "b-a 8 1 0" -@ "b-a 8 1 1" -write c/setup